streamlit web/requirements.txt
keras==2.7.0
matplotlib-inline==0.1.3
numpy==1.21.4
opencv-python==4.5.4.58
Pillow==8.4.0
pytesseract==0.3.8
streamlit==1.2.0
streamlit-cropper==0.1.3
tensorflow==2.7.0
